moves welcome to own component
